move item editor to its own "page"
